requirements.txt
requests
python-dotenv
pyyaml
pydantic